add map3, Player bubble power, count

+++ include/type.h
@@ -21,7 +21,7 @@ typedef struct
     float p2_pos_y;
 } Config;
 
-using map_t = std::vector<std::vector<uint8_t>>;
+using map_t = std::vector<std::vector<int16_t>>;
 using mapidx_t = std::tuple<uint8_t, uint8_t>;
 
 #endif // __TYPE_H_
